Reimplementing Report structure.
